move market scanning tab to individual module
